+++ tests/Eq.hs
@@ -4,18 +4,17 @@ import Data.Double()
 
 main :: IO ()
 main = do
-  putStrLn $ show [1==1, 'a'=='a', 1.1==1.1,
+  putStrLn $ show [1==(1::Int), 'a'=='a', 1.1==(1.1::Double),
                    True==True, False==False,
-                   (Nothing::Maybe Int)==Nothing, Just 1 == Just 1,
-                   [1,2,3] == [1,2,3],
-                   (1,2) == (1,2),
+                   (Nothing::Maybe Int)==Nothing, Just (1::Int) == Just 1,
+                   [1,2,3] == [1,2,3::Int],
+                   (1,2) == (1::Int,2::Int),
                    (Left 1 :: Either Int Char) == Left 1, (Right 'a' :: Either Int Char) == Right 'a'
                   ]
-  putStrLn $ show [1==2, 'a'=='b', 1.1==1.2,
+  putStrLn $ show [1==(2::Int), 'a'=='b', 1.1==(1.2::Double),
                    True==False, False==True,
-                   Nothing==Just 1, Just 1 == Nothing,
-                   [1,2,3] == [1,2,4],
-                   (1,2) == (1,4),
-                   Left 1 == Right 'a', Right 'a' == Left 1
+                   Nothing==Just (1::Int), Just (1::Int) == Nothing,
+                   [1,2,3] == [1,2,4::Int],
+                   (1,2) == (1::Int,4::Int),
+                   Left (1::Int) == Right 'a', Right 'a' == Left (1::Int)
                   ]
-
