+++ tests/Word.hs
@@ -4,10 +4,10 @@ import qualified Data.Word as W
 
 main :: IO ()
 main = do
-  putStrLn $ show 4294967295
-  putStrLn $ show (W.intToWord 1000)
+  putStrLn $ show (4294967295::Int)
+  putStrLn $ show (W.intToWord (1000::Int))
   putStrLn $ show twoTo32M1
   putStrLn $ show $ (*) twoTo32M1 twoTo32M1
 
 twoTo32M1 :: W.Word
-twoTo32M1 = W.intToWord 4294967295
+twoTo32M1 = W.intToWord (4294967295::Int)
